ingestion/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
kafka-python==2.2.12
MarkupSafe==3.0.2
mysql-connector==2.2.9
mysql-connector-python==9.3.0
PyMySQL==1.1.1
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
cryptography